quiz

+++ src/main/groovy/com/brainup/readby/util/AmazonClient.groovy
@@ -7,6 +7,7 @@ import com.amazonaws.services.s3.AmazonS3Client
 import com.amazonaws.services.s3.model.CannedAccessControlList
 import com.amazonaws.services.s3.model.DeleteObjectRequest
 import com.amazonaws.services.s3.model.PutObjectRequest
+import groovy.util.logging.Slf4j
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Service
 import org.springframework.web.multipart.MultipartFile
@@ -14,6 +15,7 @@ import org.springframework.web.multipart.MultipartFile
 import javax.annotation.PostConstruct
 
 @Service
+@Slf4j
 class AmazonClient {
 
     private AmazonS3 s3client
@@ -30,8 +32,7 @@ class AmazonClient {
     @Value('${amazonProperties.secretKey}')
     private String secretKey
 
-    @Value('${amazonProperties.childFolder}')
-    private String childFolder
+
 
     @PostConstruct
     private void initializeAmazon() {
@@ -39,13 +40,13 @@ class AmazonClient {
         this.s3client = new AmazonS3Client(credentials)
     }
 
-    public String uploadFile(MultipartFile multipartFile) {
+    public String uploadFile(MultipartFile multipartFile,String childFolder) {
         String fileUrl = ""
         try {
             File file = convertMultiPartToFile(multipartFile)
             String fileName = generateFileName(multipartFile)
             fileUrl = endpointUrl + "/" + childFolder + "/" + fileName
-            uploadFileTos3bucket(fileName, file)
+            uploadFileTos3bucket(fileName, file,childFolder)
             file.delete()
         } catch (Exception e) {
             e.printStackTrace()
@@ -65,14 +66,18 @@ class AmazonClient {
         return new Date().getTime() + "-" + multiPart.getOriginalFilename().replace(" ", "_")
     }
 
-    private void uploadFileTos3bucket(String fileName, File file) {
+    private void uploadFileTos3bucket(String fileName, File file,String childFolder) {
         s3client.putObject(new PutObjectRequest(bucketName+ "/" + childFolder, fileName, file)
                 .withCannedAcl(CannedAccessControlList.PublicRead))
     }
 
-    public String deleteFileFromS3Bucket(String fileUrl) {
-        String fileName = fileUrl.substring(fileUrl.lastIndexOf("/") + 1)
-        s3client.deleteObject(new DeleteObjectRequest(bucketName+ "/" + childFolder, fileName))
-        return "Successfully deleted"
+    public void deleteFileFromS3Bucket(String fileUrl,String childFolder) {
+        if(fileUrl != null) {
+            String fileName = fileUrl.substring(fileUrl.lastIndexOf("/") + 1)
+            s3client.deleteObject(new DeleteObjectRequest(bucketName + "/" + childFolder, fileName))
+            log.info "Successfully deleted"
+        }else{
+            log.info "fileurl is null"
+        }
     }
 }
